feat: decouple uiState handler from android viewmodel
