+++ platform/uistate/src/test/java/com/example/uistate/UiStateHandlerTest.kt
@@ -1,0 +1,9 @@
+package com.example.uistate
+
+import kotlinx.coroutines.ExperimentalCoroutinesApi
+
+@ExperimentalCoroutinesApi
+internal class UiStateHandlerTest {
+
+
+}
